--- lib/models/preferences_model/accessibility/accessibility.dart
@@ -1,14 +1,0 @@
-// import 'package:freezed_annotation/freezed_annotation.dart';
-
-// part 'accessibility.freezed.dart';
-// part 'accessibility.g.dart';
-
-// @freezed
-// class Accessibility with _$Accessibility {
-//   factory Accessibility({
-    
-//   }) = _Accessibility;
-
-//   factory Accessibility.fromJson(Map<String, dynamic> json) => _$Accessibility(json);
-//   Map<String, dynamic> toJson() => _$Accessibility(this);
-// }
