--- lib/models/preferences_model/language_and_region/language_and_region.dart
@@ -1,14 +1,0 @@
-// import 'package:freezed_annotation/freezed_annotation.dart';
-// import 'package:zc_desktop_flutter/ui/shared/preferences/preferenceswidgets/advanced/advanced_viewmodel.dart';
-
-// part 'language_and_region.freezed.dart';
-// part 'language_and_region.g.dart';
-
-// @freezed
-// class LanguageAndRegion with _$LanguageAndRegion {
-//   factory LanguageAndRegion({
-//   }) = _LanguageAndRegion;
-
-//   factory LanguageAndRegion.fromJson(Map<String, dynamic> json) => _$Advanced(json);
-//   Map<String, dynamic> toJson() => _$LanguageAndRegion(this);
-// }
