--- lib/models/preferences_model/mark_as_read/mark_as_read.dart
@@ -1,14 +1,0 @@
-// import 'package:freezed_annotation/freezed_annotation.dart';
-
-// part 'mark_as_read.freezed.dart';
-// part 'mark_as_read.g.dart';
-
-// @freezed
-// class MarkAsRead with _$MarkAsRead {
-//   factory MarkAsRead({
-    
-//   }) = _MarkAsRead;
-
-//   factory MarkAsRead.fromJson(Map<String, dynamic> json) => _$MarkAsRead(json);
-//   Map<String, dynamic> toJson() => _$MarkAsRead(this);
-// }
