+++ lib/services/users_loacal_data.dart
@@ -1,16 +1,16 @@
-import 'package:zc_desktop_flutter/models/user_model.dart';
+import 'package:zc_desktop_flutter/model/app_models.dart';
 
-final usersData = <User>[
-  User(name: 'Dan', profileImage: 'path/to/image'),
-  User(name: 'Bunmi', profileImage: 'path/to/image'),
-  User(name: 'Mark', profileImage: 'path/to/image'),
-  User(name: 'Naza', profileImage: 'path/to/image'),
-  User(name: 'Xyluz', profileImage: 'path/to/image'),
-  User(name: 'Katherine', profileImage: 'path/to/image'),
-  User(name: 'Dumebi', profileImage: 'path/to/image'),
-  User(name: 'Olarenwanju', profileImage: 'path/to/image'),
-  User(name: 'Adema', profileImage: 'path/to/image'),
-  User(name: 'Protector', profileImage: 'path/to/image'),
-  User(name: 'Demms', profileImage: 'path/to/image'),
-  User(name: 'Developer Jamiu', profileImage: 'path/to/image'),
+final usersData = <DummyUser>[
+  DummyUser(name: 'Dan', profileImage: 'path/to/image'),
+  DummyUser(name: 'Bunmi', profileImage: 'path/to/image'),
+  DummyUser(name: 'Mark', profileImage: 'path/to/image'),
+  DummyUser(name: 'Naza', profileImage: 'path/to/image'),
+  DummyUser(name: 'Xyluz', profileImage: 'path/to/image'),
+  DummyUser(name: 'Katherine', profileImage: 'path/to/image'),
+  DummyUser(name: 'Dumebi', profileImage: 'path/to/image'),
+  DummyUser(name: 'Olarenwanju', profileImage: 'path/to/image'),
+  DummyUser(name: 'Adema', profileImage: 'path/to/image'),
+  DummyUser(name: 'Protector', profileImage: 'path/to/image'),
+  DummyUser(name: 'Demms', profileImage: 'path/to/image'),
+  DummyUser(name: 'Developer Jamiu', profileImage: 'path/to/image'),
 ];
