+++ lib/ui/shared/smart_widgets/preferences/preferenceswidgets/messages_media_preference/message_media_preference_viewModel.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 
 import 'package:stacked/stacked.dart';
 import 'package:zc_desktop_flutter/app/app.locator.dart';
-import 'package:zc_desktop_flutter/models/preferences_model/messages_and_media/messages_and_media.dart';
+import 'package:zc_desktop_flutter/model/app_models.dart';
 import 'package:zc_desktop_flutter/services/local_storage_service.dart';
 
 import 'message_media_preferenceview.dart';
