+++ lib/ui/shared/smart_widgets/preferences/preferenceswidgets/sidebar/preferences_sidebar_viewmodel.dart
@@ -4,7 +4,7 @@ import 'package:stacked/stacked.dart';
 import 'package:zc_desktop_flutter/app/app.locator.dart';
 import 'package:zc_desktop_flutter/core/enums/pre_bar.dart';
 import 'package:zc_desktop_flutter/core/enums/pre_sidebar.dart';
-import 'package:zc_desktop_flutter/models/preferences_model/side_bar/sidebar.dart';
+import 'package:zc_desktop_flutter/model/app_models.dart';
 import 'package:zc_desktop_flutter/services/local_storage_service.dart';
 
 class SideBarViewModel extends BaseViewModel {
