+++ lib/ui/views/main/channels_page/channels_viewmodel.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:stacked/stacked.dart';
 import 'package:zc_desktop_flutter/app/app.locator.dart';
 import 'package:zc_desktop_flutter/app/app.logger.dart';
-import 'package:zc_desktop_flutter/models/workspace.dart';
+import 'package:zc_desktop_flutter/model/app_models.dart';
 import 'package:zc_desktop_flutter/services/channel_service/channel_service.dart';
 
 class ChannelsViewModel extends BaseViewModel {
